[frontend] end of refacto Draft widgets  (#15689)
